move pagination parameter rewrite logic from StandardParameterBuilder to PaginationParameterRewriter
